rename attribution_class to AttributionClass. Camels are cute
