+++ src/lib/attributionClasses.js
