+++ src/schema/artwork/attributionClass.js
